Fix: Various
- Fix positions in the case of draws
- Add new scoring systems
- Add developer info page
- Change navigation buttons to links for easier use

+++ backend/database/results.py
@@ -80,7 +80,7 @@ def courseResultToJSON(result, league, eventsList):
         for event in range(len(result[3].split(';'))):
             eventsByCompetitor.append({
                 'event': eventList[event],
-                'points': int(pointsList[event])
+                'points': float(pointsList[event])
             })
         # Get points for each event in order
         points = []
